ENT-11267: Introducing VerificationService, which implements VerificationSupport in terms of node-based services
